+++ packages/moengage_cards/moengage_cards_platform_interface/lib/src/model/display_control.dart
@@ -55,4 +55,9 @@ class DisplayControl {
         keyIsPinned: isPinned,
         keyShowTime: showTime.toJson()
       };
+
+  @override
+  String toString() {
+    return 'DisplayControl{expireAt: $expireAt, expireAfterSeen: $expireAfterSeen, expireAfterDelivered: $expireAfterDelivered, maxCount: $maxCount, isPinned: $isPinned, showTime: $showTime}';    
+  }
 }