+++ packages/moengage_cards/moengage_cards_platform_interface/CHANGELOG.md
@@ -1,5 +1,9 @@
 # MoEngage Cards Platform Interface
           
+# Next Release
+## Next version
+- Added Accessibility support in Cards.
+
 # 25-11-2024
 
 ## 4.0.2